test(rust): pin the map_to_local rotation and two clash contracts
Test-only. Three mutation-verified gaps in rust/core and rust/clash.

core/src/georef.rs `map_to_local` — BOTH rotation terms were unpinned:

  let x = inv_scale * (cos_r * dx + sin_r * dy);   flip -> survives 137/137
  let y = inv_scale * (-sin_r * dx + cos_r * dy);  flip -> survives 137/137

`test_georef_map_to_local` used the identity rotation (sin_r = 0), and
`test_georef_with_rotation` exercised only `local_to_map`, with y = 0 —
which zeroes the very term that would have caught it. So a sign
regression in the inverse transform was invisible from two directions at
once. This is the same shape as the Y-up/Z-up sign drop that mirrored an
entire HBJSON export earlier in this campaign.

Bounded by a control: the analogous flip in `local_to_map`'s `e` term IS
caught, by georef_matrix.rs's to_matrix cross-check — which never calls
`map_to_local`. Only the inverse path was blind. Now pinned by a
round-trip at a genuine 30-degree rotation with both local coordinates
non-zero.

clash/src/narrow.rs — the clearance rule is documented as "ANY gap
within the required clearance is a violation", but the test used
clearance 1.5 against a 1.0 gap, far past the line where `<=` and `<`
agree. `<=` -> `<` survived 22/22. Pinned at exactly the gap.

clash/src/triangle.rs — `tri_tri_distance` returns (dist, pA, pB) with
pA on triangle A. Every test discarded both points, and the only caller
feeds them to `mid()` and `bounds_of_points()`, both symmetric in their
arguments — so swapping pA/pB was invisible. Survived 22/22. LATENT: no
current consumer is asymmetric, but the contract is documented.

georef.rs was 820 lines against a 793 ratchet budget once the new test
landed, so its tests moved to a sibling georef_tests.rs via
`#[cfg(test)] #[path = ...]` — shrink, not raise. Ratchet 4/4.

clash 24 pass (was 22), core 138 (was 137).
