+++ rust/core/src/georef.rs
@@ -709,85 +709,5 @@ impl RtcOffset {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_georef_local_to_map() {
-        let mut georef = GeoReference::new();
-        georef.eastings = 500000.0;
-        georef.northings = 5000000.0;
-        georef.orthogonal_height = 100.0;
-
-        let (e, n, h) = georef.local_to_map(10.0, 20.0, 5.0);
-        assert!((e - 500010.0).abs() < 1e-10);
-        assert!((n - 5000020.0).abs() < 1e-10);
-        assert!((h - 105.0).abs() < 1e-10);
-    }
-
-    #[test]
-    fn test_georef_map_to_local() {
-        let mut georef = GeoReference::new();
-        georef.eastings = 500000.0;
-        georef.northings = 5000000.0;
-        georef.orthogonal_height = 100.0;
-
-        let (x, y, z) = georef.map_to_local(500010.0, 5000020.0, 105.0);
-        assert!((x - 10.0).abs() < 1e-10);
-        assert!((y - 20.0).abs() < 1e-10);
-        assert!((z - 5.0).abs() < 1e-10);
-    }
-
-    #[test]
-    fn test_georef_with_rotation() {
-        let mut georef = GeoReference::new();
-        georef.eastings = 0.0;
-        georef.northings = 0.0;
-        // 90 degree rotation
-        georef.x_axis_abscissa = 0.0;
-        georef.x_axis_ordinate = 1.0;
-
-        let (e, n, _) = georef.local_to_map(10.0, 0.0, 0.0);
-        // After 90 degree rotation: (10, 0) -> (0, 10)
-        assert!(e.abs() < 1e-10);
-        assert!((n - 10.0).abs() < 1e-10);
-    }
-
-    #[test]
-    fn test_rtc_offset() {
-        let positions = vec![
-            500000.0f32,
-            5000000.0,
-            0.0,
-            500010.0,
-            5000010.0,
-            10.0,
-            500020.0,
-            5000020.0,
-            20.0,
-        ];
-
-        let offset = RtcOffset::from_positions(&positions);
-        assert!(offset.is_significant());
-        assert!((offset.x - 500010.0).abs() < 1.0);
-        assert!((offset.y - 5000010.0).abs() < 1.0);
-    }
-
-    #[test]
-    fn test_rtc_apply() {
-        let mut positions = vec![500000.0f32, 5000000.0, 0.0, 500010.0, 5000010.0, 10.0];
-
-        let offset = RtcOffset {
-            x: 500000.0,
-            y: 5000000.0,
-            z: 0.0,
-        };
-
-        offset.apply(&mut positions);
-
-        assert!((positions[0] - 0.0).abs() < 1e-5);
-        assert!((positions[1] - 0.0).abs() < 1e-5);
-        assert!((positions[3] - 10.0).abs() < 1e-5);
-        assert!((positions[4] - 10.0).abs() < 1e-5);
-    }
-}
+#[path = "georef_tests.rs"]
+mod georef_tests;
